Alerts 2.0

+++ samples/server/petstore/jaxrs-resteasy/eap-java8/src/main/java/io/swagger/api/impl/StoreApiServiceImpl.java
@@ -22,7 +22,7 @@ public class StoreApiServiceImpl implements StoreApi {
       // do some magic!
       return Response.ok().build();
   }
-      public Response getOrderById(Integer orderId,SecurityContext securityContext) {
+      public Response getOrderById(Long orderId,SecurityContext securityContext) {
       // do some magic!
       return Response.ok().build();
   }
